BigInteger를 Long으로 변경

+++ src/main/java/com/example/trackify/task/repository/TaskRepository.java
@@ -5,7 +5,6 @@ import jakarta.persistence.EntityManager;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -18,7 +17,7 @@ public class TaskRepository {
         em.persist(task);
     }
 
-    public Task findOne(BigInteger taskSequence) {
+    public Task findOne(Long taskSequence) {
         return em.find(Task.class, taskSequence);
     }
 
@@ -26,7 +25,7 @@ public class TaskRepository {
         return em.createQuery("select t from Task t", Task.class).getResultList();
     }
 
-    public void update(BigInteger taskSequence, String taskTitle, LocalDate taskStartDate, LocalDate taskEndDate){
+    public void update(Long taskSequence, String taskTitle, LocalDate taskStartDate, LocalDate taskEndDate){
         em.createQuery("update Task t set t.taskTitle = :title, t.taskStartDate = :startDate, t.taskEndDate = :endDate where t.taskSequence = :sequence", Task.class)
                 .setParameter("title", taskTitle)
                 .setParameter("startDate", taskStartDate)
@@ -35,14 +34,14 @@ public class TaskRepository {
                 .executeUpdate();
     }
 
-    public void updateStatus(BigInteger taskSequence, Integer taskStatus){
+    public void updateStatus(Long taskSequence, Integer taskStatus){
         em.createQuery("update Task t set t.taskStatus = :status where t.taskSequence = :sequence", Task.class)
                 .setParameter("status", taskStatus)
                 .setParameter("sequence", taskSequence)
                 .executeUpdate();
     }
 
-    public void delete(BigInteger taskSequence){
+    public void delete(Long taskSequence){
         em.createQuery("delete from Task t where t.taskSequence = :sequence", Task.class)
                 .setParameter("sequence", taskSequence)
                 .executeUpdate();
